Don't fail when a perflib counter isn't found
